Spec 743.1 — CPU clk + alarm primitive monotonic (no 32-bit wrap)
BUG-025 root: maincpu absolute CLOCK wrapped at 2^32 (clkAdd=u32(a+b), cycles
setter u32, alarm sentinel 0xffffffff) → near the boundary an alarm wrapped below
clk and drainAlarms spun.

- util/uint.ts: CLOCK is a monotonic JS number (modern VICE CLOCK is uint64), not
  a uint32. Add CLOCK_NEVER = MAX_SAFE_INTEGER (the one disabled-clock sentinel).
  clkAdd = a + b (no u32). clkDelta keeps u32 (short non-neg delta only).
- cpu/cpu65xx-vice.ts: cycles setter stores absolute clk verbatim (no u32).
- alarm/alarm-context.ts: empty/disabled sentinel CLOCK_MAX repointed to
  CLOCK_NEVER (not 0xffffffff); capture/restore drop >>>0; header comment updated
  to monotonic semantics. Dispatch offset (short non-neg) unchanged.

Below 2^32 behaviour is identical (monotonic == u32 there); only the >2^32 wrap is
removed. CIA/VIC absolute-clock scheduling still wraps (743.2/.3) so the full
forced-boundary machine run still trips — fixed in the next slices.

Gate: npm run probe:743-1 (scripts/probe-743-1-alarm-primitive.mjs, 12/12) —
clkAdd/CPU cycles/alarm schedule/capture-restore/drain all monotonic past 2^32,
empty sentinel = CLOCK_NEVER. probe-single-path 25/25; boot + inspect clean.

+++ src/runtime/headless/cpu/cpu65xx-vice.ts
@@ -244,7 +244,8 @@ export class Cpu65xxVice {
   public get sp():    BYTE { return this.reg_sp; }
   public set sp(v: BYTE)   { this.reg_sp = u8(v); }
   public get cycles(): CLOCK { return this.clk; }
-  public set cycles(v: CLOCK) { this.clk = u32(v); }
+  // Spec 743 — absolute clk is monotonic; do NOT u32 it (a long run exceeds 2^32).
+  public set cycles(v: CLOCK) { this.clk = v; }
   /** Composite P register including flag_n/flag_z view. Compatible
    *  with legacy `flags` field semantics. */
   public get flags(): BYTE {
